refactor: uncomment markdown generation
